added testing scripts for setup and api also added dockerfiles

+++ setup.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+# check if root
 if [ "$(whoami)" != "root" ]
 then
         echo "[error] you must be root or run using sudo $0"
@@ -10,27 +11,185 @@ tui_h=20
 tui_w=75
 tui_t=6
 
-declare -A download_urls=(
-        ["LICENSE"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/LICENSE"
-        ["functions.sh"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/functions.sh"
-        ["relayctl.sh"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/relayctl.sh"
-        ["schedule.list"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/schedule.list"
-        ["scheduler.sh"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/scheduler.sh"
-        ["settings.sh"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/settings.sh"
-	["external.sh"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/external.sh"
-	["inputs.list"]="https://raw.githubusercontent.com/ReinierNel/relayctl/main/inputs.list"
+if [ "$1" = "test" ]
+then
+        testing="true"
+else
+        testing="false"
+fi
+
+if [ "$1" = "-y" ]
+then
+        yes_2_all="true"
+else
+        yes_2_all="false"
+fi
+
+# gitlab repo base url to download raw files
+download_url="https://raw.githubusercontent.com/ReinierNel/relayctl"
+
+# files in repo to download later
+declare -a download_files=(
+        "LICENSE"
+        "functions.sh"
+        "relayctl.sh"
+        "schedule.list"
+        "scheduler.sh"
+        "settings.sh"
+	"external.sh"
+	"inputs.list"
+        "api-key.sh"
 )
 
+# check if status is 0 else exit 1
 function check_exit_status() {
-        if [ ! "$exitstatus" = 0 ]
+        if [ ! "$1" = "0" ]
         then
-                echo "[warn] user canceled!"
+                echo "[warn] user canceled. on $2"
                 exit 1
         fi
 }
 
-# licence agreement
+# fetch branches on github repo
+function fetch-branches() {
+        #get_branches=($(curl --silent https://api.github.com/repos/reiniernel/relayctl/branches | jq .[]."name"))
+        
+        mapfile -t get_branches < <(curl --silent https://api.github.com/repos/reiniernel/relayctl/branches | jq -r .[]."name")
 
+        branches=()
+        for branch in "${get_branches[@]}"
+        do
+                branches+=("$branch")
+                branches+=("")
+        done
+}
+
+# download files
+function download_files() {
+        # setup folders
+        mkdir /etc/relayctl
+
+        # download needed files
+        dl_count="${#download_url[@]}"
+        dl_loading_bar_cunks=$((100 / dl_count))
+        dl_loading_bar=0
+
+        {
+                for dl_item in "${download_files[@]}"
+                do
+                        dl_loading_bar=$((dl_loading_bar + dl_loading_bar_cunks))
+                        echo "$dl_loading_bar"
+                        curl --silent "$download_url/$1/$dl_item" --output "/etc/relayctl/$dl_item"
+                        chmod +x "/etc/relayctl/$dl_item"
+                        chown root:gpio "/etc/relayctl/$dl_item"
+                        dl_loading_bar=$((dl_loading_bar + dl_loading_bar_cunks))
+                done
+        } | whiptail --gauge "Downloading files from Github..." 6 50 0
+}
+
+# update file prems and settings files
+function update_files() {
+        chown -R root:gpio /etc/relayctl
+        chmod 775 /etc/relayctl
+
+        chmod 775 /etc/relayctl/schedule.list
+        chmod 775 /etc/relayctl/inputs.list
+        chmod -x /etc/relayctl/LICENSE
+
+        # update settings.sh
+        #sed -i "s/__OUT_GPIO_PIN__/${gpio_select[*]}/g" /etc/relayctl/settings.sh
+        #sed -i "s/__IN_GPIO_PIN__/${input_gpio_select[*]}/g" /etc/relayctl/settings.sh
+        sed -i "s/__OUT_GPIO_PIN__/$1/g" /etc/relayctl/settings.sh
+        sed -i "s/__IN_GPIO_PIN__/$2/g" /etc/relayctl/settings.sh
+        sed -i "s/__SCHEDULAR_FREQUEMCY__/10/g" /etc/relayctl/settings.sh
+}
+
+# update rc.local
+function update_rc_local() {
+        cat > /tmp/rc.local <<EOF
+$(head -n -1 /etc/rc.local)
+
+/etc/relayctl/relayctl.sh test
+/etc/relayctl/scheduler.sh &
+/etc/relayctl/external.sh &
+
+exit 0
+EOF
+
+        rm -f /etc/rc.local
+        mv /tmp/rc.local /etc/rc.local
+        chmod 755 /etc/rc.local
+}
+
+# install api
+function install_api() {
+        if [ "$?" = "0" ]
+        then
+                apt-get update
+                apt-get upgrade -y
+                apt-get install -y nginx fcgiwrap jq
+
+                usermod -a -G gpio www-data
+                mkdir -p /usr/lib/cgi-bin
+
+                curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/$1/api/cgi-bin/api.cgi" --output "/usr/lib/cgi-bin/api.cgi"
+                chown www-data:gpio /usr/lib/cgi-bin/api.cgi
+                chmod +x /usr/lib/cgi-bin/api.cgi
+                rm -f /etc/nginx/sites-available/default
+                curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/$1/api/nginx/default" --output "/etc/nginx/sites-available/default"
+                curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/$1/api/nginx/fastcgi_params" --output "/etc/nginx/fastcgi_params"
+                curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/$1/api-key.sh" --output "/etc/relayctl/api-key.sh"
+                chmod 600 /etc/relayctl/api-key.sh
+                chown root:root /etc/relayctl/api-key.sh
+                cp /usr/share/doc/fcgiwrap/examples/nginx.conf /etc/nginx/fcgiwrap.conf
+                sed -i "s/user www-data/user root/g" /etc/nginx/nginx.conf
+                service nginx restart
+                api_key=$(hexdump -n 16 -e '4/4 "%08X" 1 "\n"' /dev/urandom)
+                openssl passwd -6 -salt "$(tr -dc A-Za-z0-9 </dev/urandom | head -c 13 ; echo '')" -stdin -noverify <<< "$api_key" >> /etc/relayctl/api.key
+                chown root:gpio /etc/relayctl/api.key
+        fi
+}
+
+# install everting scip whiptail
+if [ "$yes_2_all" = "true" ]
+then
+        download_files "main"
+        update_files "17 18 27 22 23 24" "25 5 6 12 13 26"
+
+fi
+
+
+# ask use to install jq if not already presint
+read -r -d '' info_msg <<'EOF'
+This installer needs the program jq to work. For more information about jq please visit https://stedolan.github.io/jq/
+
+Can we go ahead and install jq onto this Pi?
+
+Select “Yes” to continue or “No” to cancel
+EOF
+
+if ! command -v jq &> /dev/null
+then
+        whiptail --title "Setup Relayctl" --yesno "$info_msg" "$tui_h" "$tui_w" 3>&1 1>&2 2>&3
+
+        check_exit_status "$?" "Install jq?"
+
+        apt update
+        apt install -y jq
+fi
+
+# selec install branch
+read -r -d '' select_branch_msg <<'EOF'
+Select from witch github branch to install from, For stable releases select main
+EOF
+
+fetch-branches
+branch_selected=$(whiptail --title "Setup Relayctl" --menu "$select_branch_msg" "$tui_h" "$tui_w" "$tui_t" "${branches[@]}" 3>&1 1>&2 2>&3)
+
+check_exit_status "$?" "Select Branch"
+
+
+# licence agreement
 read -r -d '' licence_msg <<'EOF'
 MIT License
 
@@ -57,11 +216,9 @@ SOFTWARE.
 Do you Agree?
 EOF
 
-licence_agreement=$(whiptail --title "Setup Relayctl" --yesno "$licence_msg" 40 75 3>&1 1>&2 2>&3)
+whiptail --title "Setup Relayctl" --yesno "$licence_msg" 40 75 3>&1 1>&2 2>&3
 
-exitstatus=$?
-
-check_exit_status
+check_exit_status "$?"
 
 # relays connected?
 read -r -d '' welcome_msg <<'EOF'
@@ -77,11 +234,9 @@ This wizard will help you set up your relays on your Pi.
 Have you connected your relays to the Pi?. If not connect them now and hit OK.
 EOF
 
-relays_connected=$(whiptail --title "Setup Relayctl" --yesno "$welcome_msg" "$tui_h" "$tui_w" 3>&1 1>&2 2>&3)
+whiptail --title "Setup Relayctl" --yesno "$welcome_msg" "$tui_h" "$tui_w" 3>&1 1>&2 2>&3
 
-exitstatus=$?
-
-check_exit_status
+check_exit_status "$?" "Inro"
 
 # select relay gpio
 read -r -d '' gpio_select_msg <<'EOF'
@@ -103,9 +258,7 @@ gpio_select=$(whiptail --title "Setup Relayctl" --checklist \
 "24" "PIN 18" OFF \
 3>&1 1>&2 2>&3)
 
-exitstatus=$?
-
-check_exit_status
+check_exit_status "$?" "Relay GPIO Select"
 
 # select inputs gpio
 read -r -d '' input_gpio_select_msg <<'EOF'
@@ -127,54 +280,15 @@ input_gpio_select=$(whiptail --title "Setup Relayctl" --checklist \
 "26" "PIN 37" OFF \
 3>&1 1>&2 2>&3)
 
-exitstatus=$?
+check_exit_status "$?" "Switches GPIO Select"
 
-check_exit_status
+if [ "$testing" = "false" ]
+then
 
-# set scheduler frequency
-read -r -d '' schedule_msg1 <<'EOF'
-Enter a value in seconds how often should the scheduler check
-if a relay sould be on or off recommended 10 seconds.
-
-A lower value will increase CPU load and also make the log files
-grow faster.
-EOF
-
-schedule_frequency=$(whiptail --title "Setup Relayctl" --inputbox "$schedule_msg1" "$tui_h" "$tui_w" 3>&1 1>&2 2>&3)
-
-exitstatus=$?
-
-check_exit_status
-
-# setup folders
-mkdir /etc/relayctl
-chown root:gpio /etc/relayctl
-chmod 775 /etc/relayctl
-# download needed files
-dl_count=${#download_urls[@]}
-dl_loading_bar_cunks=$((100 / 8))
-dl_loading_bar=0
-
-{
-        for dl_item in ${!download_urls[@]}
-        do
-                dl_loading_bar=$(($dl_loading_bar + $dl_loading_bar_cunks))
-                echo "$dl_loading_bar"
-                curl --silent "${download_urls[$dl_item]}" --output "/etc/relayctl/$dl_item"
-                chmod +x "/etc/relayctl/$dl_item"
-		chown root:gpio "/etc/relayctl/$dl_item"
-                dl_loading_bar=$(($dl_loading_bar + $dl_loading_bar_cunks))
-        done
-} | whiptail --gauge "Downloading files from Github..." 6 50 0
-
-chmod 775 /etc/relayctl/schedule.list
-chmod 775 /etc/relayctl/inputs.list
-chmod -x /etc/relayctl/LICENSE
-
-# update settings.sh
-sed -i "s/__OUT_GPIO_PIN__/${gpio_select[@]}/g" /etc/relayctl/settings.sh
-sed -i "s/__IN_GPIO_PIN__/${input_gpio_select[@]}/g" /etc/relayctl/settings.sh
-sed -i "s/__SCHEDULAR_FREQUEMCY__/$schedule_frequency/g" /etc/relayctl/settings.sh
+        download_files "$branch_selected"
+        update_files "${gpio_select[*]}" "${input_gpio_select[*]}"
+        update_rc_local
+fi
 
 # testing relays
 read -r -d '' test_msg <<'EOF'
@@ -188,21 +302,6 @@ Press OK to continue with the test
 EOF
 whiptail --title "Setup Relayctl" --msgbox "$test_msg" "$tui_h" "$tui_w"
 /etc/relayctl/relayctl.sh test
-
-# update rc.local
-cat > /tmp/rc.local <<EOF
-$(head -n -1 /etc/rc.local)
-
-/etc/relayctl/relayctl.sh test
-/etc/relayctl/scheduler.sh &
-/etc/relayctl/external.sh &
-
-exit 0
-EOF
-
-rm -f /etc/rc.local
-mv /tmp/rc.local /etc/rc.local
-chmod 755 /etc/rc.local
 
 # ask sould we setup an API or not
 read -r -d '' api_msg <<'EOF'
@@ -221,34 +320,11 @@ The following software will be install
 Select Yes to install or No to continue without installing
 EOF
 
-licence_agreement=$(whiptail --title "Setup Relayctl" --yesno "$api_msg" 40 75 3>&1 1>&2 2>&3)
+install_api=$(whiptail --title "Setup Relayctl" --yesno "$api_msg" 40 75 3>&1 1>&2 2>&3)
 
-exitstatus=$?
-
-if [ "$exitstatus" = 0 ]
+if [ "$testing" = "false" ]
 then
-	apt-get update
-	apt-get upgrade -y
-	apt-get install -y nginx fcgiwrap jq
-
-	usermod -a -G gpio www-data
-	mkdir -p /usr/lib/cgi-bin
-
-	curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/main/api/cgi-bin/api.cgi" --output "/usr/lib/cgi-bin/api.cgi"
-	chown www-data:gpio /usr/lib/cgi-bin/api.cgi
-	chmod +x /usr/lib/cgi-bin/api.cgi
-	rm -f /etc/nginx/sites-available/default
-	curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/main/api/nginx/default" --output "/etc/nginx/sites-available/default"
-	curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/main/api/nginx/fastcgi_params" --output "/etc/nginx/fastcgi_params"
-	curl --silent "https://raw.githubusercontent.com/ReinierNel/relayctl/main/api-key.sh" --output "/etc/relayctl/api-key.sh"
-	chmod 600 /etc/relayctl/api-key.sh
-	chown root:root /etc/relayctl/api-key.sh
-	cp /usr/share/doc/fcgiwrap/examples/nginx.conf /etc/nginx/fcgiwrap.conf
-	sed -i "s/user www-data/user root/g" /etc/nginx/nginx.conf
-	service nginx restart
-	api_key=$(hexdump -n 16 -e '4/4 "%08X" 1 "\n"' /dev/urandom)
-	openssl passwd -6 -salt $(tr -dc A-Za-z0-9 </dev/urandom | head -c 13 ; echo '') -stdin -noverify <<< $(echo $api_key) >> /etc/relayctl/api.key
-	chown root:gpio /etc/relayctl/api.key
+        install_api "$branch_selected"
 fi
 
 # good bye
@@ -274,5 +350,23 @@ Your Pi will now reboot
 
 Press OK to reboot
 EOF
+
 whiptail --title "Setup Relayctl" --msgbox "$bye_msg" "$tui_h" "$tui_w"
-reboot
+
+if [ "$testing" = "false" ]
+then
+        reboot
+else
+        read -r -d '' test_report << EOF
+Testing Report
+
+${!testing@} = $testing
+${!branch_selected@} = $branch_selected
+${!gpio_select@} = ${gpio_select[@]}
+${!input_gpio_select@} = ${input_gpio_select[@]}
+${!install_api@} = $install_api
+EOF
+
+        whiptail --title "Setup Relayctl" --msgbox "$test_report" "$tui_h" "$tui_w"
+
+fi
